bg(routes): re-arrange routes
- re-write routes to be in separate files
- created routes/index.js to export one instance of routes

[Finishes #168976657]
