+++ package.json
@@ -24,27 +24,27 @@
     "node": "12.7.0"
   },
   "devDependencies": {
-    "@babel/core": "^7.6.2",
-    "@babel/preset-env": "^7.6.2",
-    "@babel/register": "^7.6.2",
-    "babel-watch": "^7.0.0",
-    "braces": "^3.0.2",
-    "chai": "^4.2.0",
-    "chai-http": "^4.3.0",
-    "coveralls": "^3.0.6",
-    "eslint": "^6.4.0",
-    "eslint-config-airbnb-base": "^14.0.0",
-    "eslint-plugin-import": "^2.18.2",
-    "mocha": "^6.2.0",
-    "nyc": "^14.1.1"
+    "@babel/core": "7.6.2",
+    "@babel/preset-env": "7.6.2",
+    "@babel/register": "7.6.2",
+    "babel-watch": "7.0.0",
+    "braces": "3.0.2",
+    "chai": "4.2.0",
+    "chai-http": "4.3.0",
+    "coveralls": "3.0.6",
+    "eslint": "6.4.0",
+    "eslint-config-airbnb-base": "14.0.0",
+    "eslint-plugin-import": "2.18.2",
+    "mocha": "6.2.0",
+    "nyc": "14.1.1"
   },
   "dependencies": {
-    "@babel/cli": "^7.6.2",
-    "@hapi/joi": "^16.1.4",
-    "bcrypt": "^3.0.6",
-    "dotenv": "^8.1.0",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "pg": "^7.12.1"
+    "@babel/cli": "7.6.2",
+    "@hapi/joi": "16.1.4",
+    "bcrypt": "3.0.6",
+    "dotenv": "8.1.0",
+    "express": "4.17.1",
+    "jsonwebtoken": "8.5.1",
+    "pg": "7.12.1"
   }
 }
