+++ server/routes/auth.js
@@ -1,0 +1,10 @@
+import express from 'express';
+import auth from '../controllers/auth';
+
+const router = express.Router();
+router.use(express.json());
+
+router.post('/signup', auth.signup);
+router.post('/signin', auth.signin);
+
+export default router;
